Refactor balanceCheck into own file
* 100% test coverage of asset and collectible balance check
