+++ src/__tests__/utils.test.ts
@@ -1,9 +1,7 @@
 import { BigNumber } from "bignumber.js";
 import { expect } from "chai";
 
-import { AssetTransfer } from "../parser/csvParser";
-import { testData } from "../test/util";
-import { fromWei, toWei, TEN, ONE, ZERO, assetTransfersToSummary } from "../utils";
+import { fromWei, toWei, TEN, ONE, ZERO } from "../utils";
 
 // TODO - this is super ugly at the moment and is probably missing some stuff.
 describe("toWei()", () => {
@@ -38,196 +36,5 @@ describe("fromWei()", () => {
     expect(fromWei(oneETH, 18).toFixed()).to.be.equal("1");
     expect(fromWei(oneETH, 19).toFixed()).to.be.equal("0.1");
     expect(fromWei(oneETH, 20).toFixed()).to.be.equal("0.01");
-  });
-});
-
-describe("transferToSummary()", () => {
-  it("works for integer native currency", () => {
-    const transfers: AssetTransfer[] = [
-      {
-        token_type: "native",
-        tokenAddress: null,
-        amount: new BigNumber(1),
-        receiver: testData.addresses.receiver1,
-        decimals: 18,
-        symbol: "ETH",
-        receiverEnsName: null,
-      },
-      {
-        token_type: "native",
-        tokenAddress: null,
-        amount: new BigNumber(2),
-        receiver: testData.addresses.receiver2,
-        decimals: 18,
-        symbol: "ETH",
-        receiverEnsName: null,
-      },
-      {
-        token_type: "native",
-        tokenAddress: null,
-        amount: new BigNumber(3),
-        receiver: testData.addresses.receiver3,
-        decimals: 18,
-        symbol: "ETH",
-        receiverEnsName: null,
-      },
-    ];
-    const summary = assetTransfersToSummary(transfers);
-    expect(summary.get(null)?.amount.toFixed()).to.equal("6");
-  });
-
-  it("works for decimals in native currency", () => {
-    const transfers: AssetTransfer[] = [
-      {
-        token_type: "native",
-        tokenAddress: null,
-        amount: new BigNumber(0.1),
-        receiver: testData.addresses.receiver1,
-        decimals: 18,
-        symbol: "ETH",
-        receiverEnsName: null,
-      },
-      {
-        token_type: "native",
-        tokenAddress: null,
-        amount: new BigNumber(0.01),
-        receiver: testData.addresses.receiver2,
-        decimals: 18,
-        symbol: "ETH",
-        receiverEnsName: null,
-      },
-      {
-        token_type: "native",
-        tokenAddress: null,
-        amount: new BigNumber(0.001),
-        receiver: testData.addresses.receiver3,
-        decimals: 18,
-        symbol: "ETH",
-        receiverEnsName: null,
-      },
-    ];
-    const summary = assetTransfersToSummary(transfers);
-    expect(summary.get(null)?.amount.toFixed()).to.equal("0.111");
-  });
-
-  it("works for decimals in erc20", () => {
-    const transfers: AssetTransfer[] = [
-      {
-        token_type: "erc20",
-        tokenAddress: testData.unlistedERC20Token.address,
-        amount: new BigNumber(0.1),
-        receiver: testData.addresses.receiver1,
-        decimals: 18,
-        symbol: "ULT",
-        receiverEnsName: null,
-      },
-      {
-        token_type: "erc20",
-        tokenAddress: testData.unlistedERC20Token.address,
-        amount: new BigNumber(0.01),
-        receiver: testData.addresses.receiver2,
-        decimals: 18,
-        symbol: "ULT",
-        receiverEnsName: null,
-      },
-      {
-        token_type: "erc20",
-        tokenAddress: testData.unlistedERC20Token.address,
-        amount: new BigNumber(0.001),
-        receiver: testData.addresses.receiver3,
-        decimals: 18,
-        symbol: "ULT",
-        receiverEnsName: null,
-      },
-    ];
-    const summary = assetTransfersToSummary(transfers);
-    expect(summary.get(testData.unlistedERC20Token.address)?.amount.toFixed()).to.equal("0.111");
-  });
-
-  it("works for integer in erc20", () => {
-    const transfers: AssetTransfer[] = [
-      {
-        token_type: "erc20",
-        tokenAddress: testData.unlistedERC20Token.address,
-        amount: new BigNumber(1),
-        receiver: testData.addresses.receiver1,
-        decimals: 18,
-        symbol: "ULT",
-        receiverEnsName: null,
-      },
-      {
-        token_type: "erc20",
-        tokenAddress: testData.unlistedERC20Token.address,
-        amount: new BigNumber(2),
-        receiver: testData.addresses.receiver2,
-        decimals: 18,
-        symbol: "ULT",
-        receiverEnsName: null,
-      },
-      {
-        token_type: "erc20",
-        tokenAddress: testData.unlistedERC20Token.address,
-        amount: new BigNumber(3),
-        receiver: testData.addresses.receiver3,
-        decimals: 18,
-        symbol: "ULT",
-        receiverEnsName: null,
-      },
-    ];
-    const summary = assetTransfersToSummary(transfers);
-    expect(summary.get(testData.unlistedERC20Token.address)?.amount.toFixed()).to.equal("6");
-  });
-
-  it("works for mixed payments", () => {
-    const transfers: AssetTransfer[] = [
-      {
-        token_type: "erc20",
-        tokenAddress: testData.unlistedERC20Token.address,
-        amount: new BigNumber(1.1),
-        receiver: testData.addresses.receiver1,
-        decimals: 18,
-        symbol: "ULT",
-        receiverEnsName: null,
-      },
-      {
-        token_type: "erc20",
-        tokenAddress: testData.unlistedERC20Token.address,
-        amount: new BigNumber(2),
-        receiver: testData.addresses.receiver2,
-        decimals: 18,
-        symbol: "ULT",
-        receiverEnsName: null,
-      },
-      {
-        token_type: "erc20",
-        tokenAddress: testData.unlistedERC20Token.address,
-        amount: new BigNumber(3.3),
-        receiver: testData.addresses.receiver3,
-        decimals: 18,
-        symbol: "ULT",
-        receiverEnsName: null,
-      },
-      {
-        token_type: "native",
-        tokenAddress: null,
-        amount: new BigNumber(3),
-        receiver: testData.addresses.receiver1,
-        decimals: 18,
-        symbol: "ETH",
-        receiverEnsName: null,
-      },
-      {
-        token_type: "native",
-        tokenAddress: null,
-        amount: new BigNumber(0.33),
-        receiver: testData.addresses.receiver1,
-        decimals: 18,
-        symbol: "ETH",
-        receiverEnsName: null,
-      },
-    ];
-    const summary = assetTransfersToSummary(transfers);
-    expect(summary.get(testData.unlistedERC20Token.address)?.amount.toFixed()).to.equal("6.4");
-    expect(summary.get(null)?.amount.toFixed()).to.equal("3.33");
   });
 });
